% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{complexity}
\alias{complexity}
\title{Item Complexity}
\usage{
complexity(x)
}
\arguments{
\item{x}{Vector of factor loadings}
}
\value{
Returns the item complexity for each of the items
}
\description{
Calculates item complexity for based on the factor loadings for each item
}
\examples{
complextity(x)

}
