% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcci.R
\name{mcci}
\alias{mcci}
\title{Monte Carlo Confidence Intervals for Indirect Effects}
\usage{
mcci(b1, se1, b2, se2, seed = 1234, digits = 4)
}
\arguments{
\item{b1}{First coefficient estimate}

\item{se1}{Standard error for b1}

\item{b2}{Second coefficient estimate}

\item{se2}{Standard error for b2}

\item{seed}{Set seed to replicate}

\item{digits}{Digits to round estimates}
}
\value{
Returns table of indirect effect estimates in a confidence interval of 2.5 - 97.5\%
}
\description{
Provides indirect estimates using Monte Carlo confidence intervals between 2.5 - 97.5\%
}
\note{
Monte Carlo Confidence Intervals for Indirect Effects by Louis M. Rocconi (2021). Based on Darlington & Hayes (2017, p. 463) sample syntax for Stata.
}
\examples{

mcci(.3, .21, .3, .21)
}
