% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onefacetstats.R
\name{onefacet.stats}
\alias{onefacet.stats}
\title{One Facet Statistics}
\usage{
onefacet.stats(items = c(1:10), gstudy, round = 2)
}
\arguments{
\item{items}{Number of items (defaults to presenting 1-10 items if not otherwise specified)}

\item{gstudy}{The saved gstudy object produced from using the gstudy() function}

\item{round}{Decimals to round coefficients to}
}
\value{
Returns printed results and plot containing the generalizability coefficient, relative error variance, dependability coefficient, and absolute error variance.
}
\description{
The following function uses the formulas presented in Huebner & Lucht (2019) to create a table and plot of the dependability and generalizability coefficients for various item types for a one facet design (note this will not work for two facet designs, there is a separate function for that)
}
\note{
Two reliability-like coefficients are produced: Generalizability coefficient and a Dependability Coefficient. The Generalizability coefficient is useful when making relative decisions such as rank-ordering subjects. The Dependability coefficient is useful when making absolute decision such as a subjects standing on a construct (e.g., a score of 109 on an IQ test) or when a subject must achieve a particular standard (e.g., a pass/fail exam, licensure exam).
}
\examples{
onefacet.stats(items=1:25, gstudy=g.compFear)

}
