% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolR2.R
\name{poolR2}
\alias{poolR2}
\title{Pool R-square}
\usage{
poolR2(x)
}
\arguments{
\item{x}{An imputation model object produced by \code{mice::mice}}
}
\value{
Returns the summary statistics (i.e., average, median, minimum, maximum) for R-square and adjusted R-square across all imputations
}
\description{
Returns summary statistics for R-square and adjusted R-squared for a specified model
}
\examples{

poolR2(imp1)

}
