% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RE.R
\name{RE}
\alias{RE}
\title{Relative Efficiency}
\usage{
RE(x)
}
\arguments{
\item{x}{Model object}
}
\value{
Returns the Relative Efficiency (RE) statistic for the variables included in the model
}
\description{
Calculates the Relative Efficiency (RE). These values indicate the quality (efficiency) of the estimates. Values closer to 1 indicate high quality.
}
\note{
formula 10.4 Heymans, M. W. (2019). Applied Missing Data Analysis with SPSS and (R)Studio. https://bookdown.org/mwheymans/bookmi/measures-of-missing-data-information.html

The close to 1 the better. It’s based on the fraction of missing information and the number of imputation, so the larger the amount of missing data, the greater number of imputation needed. RE = 1/(1+(FMI/m)).
}
\examples{
RE(model)

}
