% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{reshape}
\alias{reshape}
\title{GTheory Reshape}
\usage{
reshape(data)
}
\arguments{
\item{data}{Dataframe}
}
\value{
Returns data frame in long format
}
\description{
In order to run a Gtheory analysis, we need our data in long format. To do this, I'm using the melt function from the reshape2 package. You'll need to install the reshape2 package in order for this function to work.
}
\examples{
reshape(data)
}
