% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureMatrix.R
\name{structureMatrix}
\alias{structureMatrix}
\title{Structure Maxtrix}
\usage{
structureMatrix(x, cut = 0.32)
}
\arguments{
\item{x}{An object produced using the \code{psych::fa} function.}

\item{cut}{}
}
\value{
A matrix containing
}
\description{
Structure Maxtrix
}
\examples{

fa1 <- psych::fa(saq)

structureMatrix(fa1)

}
