% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{theta}
\alias{theta}
\title{IRT Theta}
\usage{
theta(theta, params)
}
\arguments{
\item{theta}{Latent ability value}

\item{params}{Object of saved IRT parameter estimates by using coef() function}
}
\value{
Returns table of likelihood of someone with that ability level (theta) would answer the item correctly (column z).
}
\description{
To evaluate the IRT model at certain values of theta, you can use the function I wrote below. You will only need to run the function once then you can call it using the command theta. Give the function the level of the latent trait you want to evaluate (theta) and where the IRT parameters are stored (params). Column z in the output gives you the likelihood of someone with that ability level (theta) would answer the item correctly.
}
\examples{
theta(0, rasch.params$items)
}
