% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twofacetstats.R
\name{twofacet.stats}
\alias{twofacet.stats}
\title{Two Facet Statistics}
\usage{
twofacet.stats(n_i = 1, n_o = 1, gstudy, round = 2)
}
\arguments{
\item{n_i}{Input number for first facet}

\item{n_o}{Input number for second facet}

\item{gstudy}{The saved gstudy object produced from using the gstudy() function}

\item{round}{Decimals to round coefficients to}
}
\value{
Returns table with relative error variance, absolute error variance, generalizability coefficient, and dependability coefficient
}
\description{
Function to calculate Generalizability and Dependability Coefficients for different numbers of designs
}
\examples{
twofacet.stats(n_i=1, n_o=1, gstudy, round=2)

}
