% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanMatrix.R
\name{dropMatrix}
\alias{dropMatrix}
\title{Drop Matrix}
\usage{
dropMatrix(matrix, vars)
}
\arguments{
\item{matrix}{Full matrix to be cleaned}

\item{vars}{Variables to add/remove from matrix}
}
\value{
Returns a matrix with variables added or removed
}
\description{
Drops/keeps variable names within the correlation matrix
}
\examples{
dropMatrix(matrix, c("var1", "var2"))

}
